# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Economia Com História API",
        "description": "Documentação completa da API para a plataforma de gamificação educacional Economia Com História. Utiliza autenticação via Bearer Token (Sanctum).",
        "contact": {
            "email": "suporte@economiacomhistoria.ao"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor API principal"
        }
    ],
    "paths": {
...
        "/admin/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Listar todas as configurações da plataforma (Apenas Admin)",
                "operationId": "indexSettings",
                "responses": {
                    "200": {
                        "description": "Configurações obtidas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Setting"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            }
        },
        "/admin/settings/{key}": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Visualizar uma configuração específica (Apenas Admin)",
                "operationId": "showSetting",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Chave da configuração",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuração obtida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Setting"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Configuração não encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Settings"
                ],
                "summary": "Atualizar o valor de uma configuração (Apenas Admin)",
                "operationId": "updateSetting",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Chave da configuração",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string",
                                        "example": "Novo Nome do Site"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuração atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Setting updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Setting"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Configuração não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "session_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Setting": {
                "title": "Setting",
                "description": "Configuração global da plataforma",
                "required": [
                    "id",
                    "key",
                    "value",
                    "type",
                    "group",
                    "is_public"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019f21a6-c469-70fd-8cff-bef624fdcf28"
                    },
                    "key": {
                        "type": "string",
                        "example": "site_name"
                    },
                    "value": {
                        "type": "string",
                        "example": "Economia com História"
                    },
                    "type": {
                        "type": "string",
                        "example": "string",
                        "enum": [
                            "string",
                            "integer",
                            "boolean",
                            "json",
                            "float"
                        ]
                    },
                    "group": {
                        "type": "string",
                        "example": "general"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nome oficial da plataforma"
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Token de autenticação Sanctum. Envie no cabeçalho: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "session_token": {
                "type": "apiKey",
                "description": "Token de sessão enviado no cabeçalho X-Session-Token",
                "name": "X-Session-Token",
                "in": "header"
            }
        }
    },
...
}